Moving some components to new files and refactor custom fields filter
