+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/catchAllLabel.tsx
@@ -1,0 +1,1 @@
+export const catchAllLabel = 'Not Applied';
