+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/familyLastName.tsx
@@ -1,0 +1,9 @@
+import { CombinedFamilyInfo } from '../../GeneratedClient';
+
+export function familyLastName(family: CombinedFamilyInfo) {
+  return (
+    family.family!.adults?.filter(
+      (adult) => family.family!.primaryFamilyContactPersonId === adult.item1?.id
+    )[0]?.item1?.lastName || '⚠ MISSING PRIMARY CONTACT'
+  );
+}
