+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/filterOption.tsx
@@ -1,0 +1,8 @@
+import { filterType } from './filterType';
+
+export type filterOption = {
+  key: string;
+  value: string | undefined;
+  type?: filterType | undefined;
+  selected: boolean;
+};
