+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/filterType.tsx
@@ -1,0 +1,6 @@
+//#region Role/Status Selection code
+
+export enum filterType {
+  Family = 1,
+  Individual = 2,
+}
