+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/getOptionValueFromSelection.tsx
@@ -1,0 +1,8 @@
+export function getOptionValueFromSelection(
+  allOptionKeysPlusSelectedOptionValue: string | string[]
+) {
+  const optionValue = [...allOptionKeysPlusSelectedOptionValue].filter(
+    (s: string) => !isNaN(Number(s))
+  )[0];
+  return optionValue ? optionValue : undefined;
+}
