+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/getUpdatedFilters.tsx
@@ -1,0 +1,12 @@
+import { filterOption } from './filterOption';
+
+export function getUpdatedFilters(
+  filters: filterOption[],
+  optionToUpdate: filterOption
+) {
+  return filters.map((filter) =>
+    filter.key === optionToUpdate.key
+      ? { ...filter, selected: !filter.selected }
+      : filter
+  );
+}
