+++ src/caretogether-pwa/src/Volunteers/VolunteerApproval/simplify.tsx
@@ -1,0 +1,7 @@
+export function simplify(input: string) {
+  // Strip out common punctuation elements and excessive whitespace, and convert to lowercase
+  return input
+    .replace(/[.,/#!$%^&*;:{}=\-_`'"'‘’‚‛“”„‟′‵″‶`´~()]/g, '')
+    .replace(/\s{2,}/g, ' ')
+    .toLowerCase();
+}
